refactor: create <IconColumnSection />
